# Add ManagedCluster informer to improve performance