--- a/pkg/apis/cluster/v1alpha1/clustergateway_health.go
+++ b/pkg/apis/cluster/v1alpha1/clustergateway_health.go
@@ -45,6 +45,10 @@ func (in *ClusterGatewayHealth) Get(ctx context.Context, name string, options *m
 }
 
 func (in *ClusterGatewayHealth) Update(ctx context.Context, name string, objInfo rest.UpdatedObjectInfo, createValidation rest.ValidateObjectFunc, updateValidation rest.ValidateObjectUpdateFunc, forceAllowCreate bool, options *metav1.UpdateOptions) (runtime.Object, bool, error) {
+	if singleton.GetSecretControl() == nil {
+		return nil, false, fmt.Errorf("loopback clients are not inited")
+	}
+
 	latestSecret, err := singleton.GetSecretControl().Get(ctx, name)
 	if err != nil {
 		return nil, false, err