--- a/pkg/featuregates/featue_gate.go
+++ b/pkg/featuregates/featue_gate.go
@@ -38,10 +38,18 @@ const (
 	// ClientIdentityPenetration enforce impersonate as the original request user
 	// when accessing apiserver in ManagedCluster
 	ClientIdentityPenetration featuregate.Feature = "ClientIdentityPenetration"
+
+	// owner: @ivan-cai
+	// beta: v1.6.0
+	//
+	// SecretCache runs a OCM ManagedCluster informer inside the apiserver which
+	// provides a cache for reading ManagedCluster.
+	OCMClusterCache featuregate.Feature = "OCMClusterCache"
 )
 
 var DefaultKubeFedFeatureGates = map[featuregate.Feature]featuregate.FeatureSpec{
 	HealthinessCheck:          {Default: false, PreRelease: featuregate.Alpha},
 	SecretCache:               {Default: true, PreRelease: featuregate.Beta},
 	ClientIdentityPenetration: {Default: false, PreRelease: featuregate.Alpha},
+	OCMClusterCache:           {Default: true, PreRelease: featuregate.Beta},
 }